refactor #79 디렉터리명 복수형으로 변경
- from `view` to `views`
- from `view_model` to `view_models`
