+++ lib/features/chat/bindings/debate_binding.dart
@@ -1,7 +1,7 @@
 import 'package:debateseason_frontend_v1/core/network/dio_client.dart';
 import 'package:debateseason_frontend_v1/features/chat/data/data_sources/room_data_source.dart';
 import 'package:debateseason_frontend_v1/features/chat/data/data_sources/vote_data_source.dart';
-import 'package:debateseason_frontend_v1/features/chat/presentation/view_model/debate_room_view_model.dart';
+import 'package:debateseason_frontend_v1/features/chat/presentation/view_models/debate_room_view_model.dart';
 import 'package:get/get.dart';
 
 class DebateBinding extends Bindings {
