+++ lib/features/chat/presentation/view_models/chat_room_view_model.dart
